2017.4.22

+++ WebRoot/administratorAccountManager.jsp
@@ -9,6 +9,8 @@
             + path + "/";
     List administrators = (List) request.getAttribute("administrators");
     List<UnitType> unitTypes=new OperationData().getUnitTypes();
+    Administrator administrator=(Administrator)request.getSession().getAttribute("adminInfo");
+    
 %>
 
 <!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN">
@@ -41,15 +43,43 @@
 	-->
 <script type="text/javascript">
 function deleteModel(id){
-    $("#deleteModel-Hidden-Id").val(id);
-    $('#deleteModal').modal('show');
-    $('#deleteModal').draggable();
+    if(id=='1'){
+        alert("唯一超级用户，不能删除");
+    }else{
+        $("#deleteModel-Hidden-Id").val(id);
+        $('#deleteModal').modal('show');
+        $('#deleteModal').draggable();
+    }  
 }
  
 function updateModel(id){
+  
+    if(id=='1'){
+        $("#updateLevel").hide();
+        $("#updateLevelLabel").hide();
+        $("#updateScope").hide();
+        $("#updateScopeLabel").hide();
+        $("#updateEmailLabel").show();
+        $("#updateEmail").show();
+        $("#updateVerification").show();
+        $("#updateVerificationLabel").show();
+        $("#example-order-button-update-verification").show();
+    }else{
+       $("#updateLevel").show();
+        $("#updateLevelLabel").show();
+        $("#updateScope").show();
+        $("#udpateScopeLabel").show();
+       /*  $("#updateEmailLabel").hide();
+        $("#updateEmail").hide();
+        $("#updateVerification").hide();
+        $("#updateVerificationLabel").hide();
+        $("#example-order-button-update-verification").hide(); */
+        
+    }
     $("#updateModel-Hidden-Id").val(id);
     $('#updateModal').modal('show');
     $('#updateModal').draggable();
+   
 }
  
 </script>
@@ -74,12 +104,14 @@ transition: opacity 0.3s linear, top 0.3s ease-out;
   </tr>
  </table>
  <br>
+ <%if(administrator.getLevel()==0){ %>
     <p align="center"  class="operation" >
         <a href="javascript:void(0)"
             class="administratorAccountManager_insert"
         >添加新审核员</a>
     </p>
      <br>
+     <%} %>
     <table class="table_content">
         <tr class="tr_one">
             <td width="150" align="center">审核员编号</td>
@@ -102,7 +134,7 @@ transition: opacity 0.3s linear, top 0.3s ease-out;
                <%--  <c:if test="${a.getScope()==0}" >
                     <td width="150" align="center">所有</td>
                </c:if> --%><!-- 此段代码不起作用 -->
-                 <c:if test="${a.getLevell()!='0' }" >
+               <%--   <c:if test="${a.getLevell()!='0' }" > --%>
                     <td width="150" align="center">
                
                   <select id="example-multiple-selected<%=a.getId()%>"  style="width:100px">
@@ -116,7 +148,7 @@ transition: opacity 0.3s linear, top 0.3s ease-out;
                            <%} %>
                 </select>
                 </td>
-                </c:if>
+               <%--  </c:if> --%>
             
                 <td bgcolor="#D1EEEE" align="center">
                     <table width="200">
@@ -125,11 +157,12 @@ transition: opacity 0.3s linear, top 0.3s ease-out;
                                 href="javascript:void(0)" onclick="updateModel('<%=a.getId()%>')"
                                 class="administratorAccountManager_update" 
                             >修改</a></td>
-
+                     <%if(administrator.getLevel()==0){ %>
                             <td width="100" align="center"><a
                                 href=" javascript:void(0)" onclick="deleteModel('<%=a.getId() %>')"
                                 class="administratorAccountManager_delete"
                             >删除 </a></td>
+                            <%} %>
                         </tr>
                     </table>
                 </td>
@@ -226,13 +259,23 @@ transition: opacity 0.3s linear, top 0.3s ease-out;
                     <input type="text" class="form-control" id="updatePassword"
                            placeholder="请输入密码" onclick="$(this).focus()">
                 </div><br>
-                <label class="col-sm-3 col-md-3   col-lg-3  control-label"  style="display:inline">级别</label>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+                <label class="col-sm-3 col-md-3   col-lg-3  control-label"  id="updateLevelLabel" style="display:inline">级别</label>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
                 <div class="col-sm-9  col-md-9  col-lg-9"   style="display:inline">
                     <input type="text" class="form-control" id="updateLevel"
                            placeholder="请输入级别" onclick="$(this).focus()">
                 </div><br>
-                  <label class="col-sm-3 col-md-3  col-lg-3   control-label"  style="display:inline">范围</label>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                <div class="col-sm-9 col-md-9  col-lg-9"  style="display:inline">
+                 <label class="col-sm-3 col-md-3   col-lg-3  control-label"  id="updateEmailLabel" style="display:inline">邮箱</label>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+                <div class="col-sm-9  col-md-9  col-lg-9"   style="display:inline">
+                    <input type="text" class="form-control" id="updateEmail"
+                           placeholder="请输入邮箱" onclick="$(this).focus()">
+                </div><br>
+                <label class="col-sm-3 col-md-3   col-lg-3  control-label"  id="updateVerificationLabel" style="display:inline">验证码</label>&nbsp;&nbsp;&nbsp;&nbsp;
+                <div class="col-sm-9  col-md-9  col-lg-9"   style="display:inline">
+                    <input type="text" class="form-control" id="updateVerification"
+                           placeholder="请输入邮箱验证码" onclick="$(this).focus()">
+                </div><br>
+                  <label class="col-sm-3 col-md-3  col-lg-3   control-label" id="updateScopeLabel" style="display:inline">范围</label>&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+                <div class="col-sm-9 col-md-9  col-lg-9"  id="updateScope" style="display:inline">
                     <select id="example-multiple-selected-update" multiple="multiple"  style="width:300px" >
                 <%
                 
@@ -251,6 +294,8 @@ transition: opacity 0.3s linear, top 0.3s ease-out;
                     <button type="button" class="btn btn-default"
                         data-dismiss="modal"
                     >关闭</button>
+                    <button type="button" class="btn btn-primary" id="example-order-button-update-verification">
+                       获取验证码</button>
                     <button type="button" class="btn btn-primary" id="example-order-button-update">
                         提交更改</button>
                 </div>
